refactor(workspace): move API compatibility tests from benchmark to unit test file
The API compatibility tests (YKeyValue vs YKeyValueLww surface parity) are
behavioral correctness tests, not performance benchmarks. Move them to
y-keyvalue-lww.test.ts where they belong alongside the other LWW unit tests.
