+++ packages/workspace/src/__benchmarks__/conflict-resolution.bench.test.ts
@@ -208,65 +208,3 @@ describe('storage comparison', () => {
 		expect(size2).toBeLessThan(size1 * 2); // But not dramatically larger
 	});
 });
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// API Compatibility
-// ═══════════════════════════════════════════════════════════════════════════════
-
-describe('API compatibility', () => {
-	test('both implementations have identical API surface', () => {
-		const { kv: kv1 } = createKv<string>();
-		const { kv: kv2 } = createKvLww<string>();
-
-		// Same methods exist
-		expect(typeof kv1.get).toBe('function');
-		expect(typeof kv2.get).toBe('function');
-
-		expect(typeof kv1.set).toBe('function');
-		expect(typeof kv2.set).toBe('function');
-
-		expect(typeof kv1.delete).toBe('function');
-		expect(typeof kv2.delete).toBe('function');
-
-		expect(typeof kv1.has).toBe('function');
-		expect(typeof kv2.has).toBe('function');
-
-		expect(typeof kv1.observe).toBe('function');
-		expect(typeof kv2.observe).toBe('function');
-
-		expect(typeof kv1.unobserve).toBe('function');
-		expect(typeof kv2.unobserve).toBe('function');
-
-		// Same properties
-		expect(kv1.map).toBeInstanceOf(Map);
-		expect(kv2.map).toBeInstanceOf(Map);
-
-		expect(kv1.yarray).toBeDefined();
-		expect(kv2.yarray).toBeDefined();
-
-		expect(kv1.doc).toBeDefined();
-		expect(kv2.doc).toBeDefined();
-	});
-
-	test('both implementations satisfy the same consumer contract', () => {
-		// Code that works with YKeyValue should work with YKeyValueLww
-		function useKV(kv: {
-			get: (k: string) => string | undefined;
-			set: (k: string, v: string) => void;
-			has: (k: string) => boolean;
-		}) {
-			kv.set('a', '1');
-			kv.set('b', '2');
-			expect(kv.get('a')).toBe('1');
-			expect(kv.has('b')).toBe(true);
-			kv.set('a', '3');
-			expect(kv.get('a')).toBe('3');
-		}
-
-		const { kv: kv1 } = createKv<string>();
-		const { kv: kv2 } = createKvLww<string>();
-
-		useKV(kv1);
-		useKV(kv2);
-	});
-});
